Break up large CostInputForm component into smaller focused components
- Split 396-line CostInputForm.tsx into smaller, single-responsibility components:
  - BasicInfoSection.tsx (192 lines): Handles year, coverage, age, tax rate, HSA/FSA contributions
  - HealthcareCategoriesSection.tsx (144 lines): Manages healthcare category estimates
  - OtherCostsSection.tsx (67 lines): Handles other healthcare costs input
  - CostInputForm.tsx (44 lines): Now orchestrates the three sections

- Each component is focused on a specific part of the form
- Maintains all existing functionality and state management
- Improves code maintainability and reusability
- All tests pass and build succeeds
